🛠 refactor: 폴더 구조 정리
3계층 구조: router & controller, service, model 계층으로 분리
