+++ src/app-error.ts
@@ -1,0 +1,10 @@
+export class AppError extends Error {
+  status: number;
+
+  constructor(name: string, httpCode: number, description: string) {
+    super(description);
+
+    this.name = name;
+    this.status = httpCode;
+  }
+}
