+++ src/aws-s3.js
