--- src/configs/index.js
@@ -1,2 +1,0 @@
-export * from './aws-s3';
-export * from './db';
