+++ src/routers/middlewares/cleanCache.js
