+++ src/session-data.d.ts
@@ -1,0 +1,9 @@
+interface User {
+  name: string;
+}
+
+export declare module 'express-session' {
+  interface SessionData {
+    user: User;
+  }
+}
